move the common function to kernel/funcs/sparse/
